+++ src/pystra/distributions/__init__.py
@@ -16,10 +16,10 @@ from .typeiiismallestvalue import Type3SmallestValue
 from .gumbel import Gumbel
 from .weibull import Weibull
 from .maximum import Maximum
-from .scipydist import ScipyDistribution
+from .scipydist import ScipyDist
 from .parent import MaxParent
 from .zeroinflated import ZeroInflated
-from .gev import GevMax, GevMin
+from .gev import GEVmax, GEVmin
 from .shiftedlognormal import ShiftedLognormal
 
 __all__ = [
@@ -41,10 +41,10 @@ __all__ = [
     "Gumbel",
     "Weibull",
     "Maximum",
-    "ScipyDistribution",
+    "ScipyDist",
     "MaxParent",
     "ZeroInflated",
-    "GevMax",
-    "GevMin",
+    "GEVmax",
+    "GEVmin",
     "ShiftedLognormal",
 ]